Added active,deactive bots,medium term and long term

+++ lib/features/authentication/data/model/user_login_model.dart
@@ -93,7 +93,7 @@ class User {
   dynamic fcmToken;
   dynamic imageProfile;
   dynamic imagePayment;
-  int? fees;
+  double? fees;
   dynamic otp;
   int? numberOfUser;
   dynamic exchange;
